backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy>=1.26.0,<2.0.0
torch>=2.1.0
torchvision>=0.16.0
captum>=0.6.0
opencv-python>=4.8.0
python-dotenv==1.0.0
supabase==2.0.3
httpx>=0.24.0,<0.25.0
pydantic==2.5.0

